+++ span.go
@@ -53,10 +53,17 @@ func (tx *Transaction) StartSpanOptions(name, spanType string, opts SpanOptions)
 	if opts.Parent == (TraceContext{}) {
 		opts.Parent = tx.traceContext
 	}
+	// Calculate the span time relative to the transaction timestamp so
+	// that wall-clock adjustments occurring after the transaction start
+	// don't affect the span timestamp.
+	if opts.Start.IsZero() {
+		opts.Start = tx.timestamp.Add(time.Since(tx.timestamp))
+	} else {
+		opts.Start = tx.timestamp.Add(opts.Start.Sub(tx.timestamp))
+	}
 	span := tx.tracer.startSpan(name, spanType, transactionID, opts)
 	binary.LittleEndian.PutUint64(span.traceContext.Span[:], tx.rand.Uint64())
 	span.stackFramesMinDuration = tx.spanFramesMinDuration
-	span.transactionTimestamp = tx.timestamp
 	tx.spansCreated++
 	tx.mu.Unlock()
 	return span
@@ -71,7 +78,7 @@ func (tx *Transaction) StartSpanOptions(name, spanType string, opts SpanOptions)
 // containing transaction's End method has been called. Spans created in this
 // way will not have the "max spans" configuration applied, nor will they be
 // considered in any transaction's span count.
-func (t *Tracer) StartSpan(name, spanType string, transactionID SpanID, transactionTimestamp time.Time, opts SpanOptions) *Span {
+func (t *Tracer) StartSpan(name, spanType string, transactionID SpanID, opts SpanOptions) *Span {
 	if opts.Parent.Trace.Validate() != nil || opts.Parent.Span.Validate() != nil || transactionID.Validate() != nil {
 		return newDroppedSpan()
 	}
@@ -82,9 +89,11 @@ func (t *Tracer) StartSpan(name, spanType string, transactionID SpanID, transact
 	if _, err := cryptorand.Read(spanID[:]); err != nil {
 		return newDroppedSpan()
 	}
+	if opts.Start.IsZero() {
+		opts.Start = time.Now()
+	}
 	span := t.startSpan(name, spanType, transactionID, opts)
 	span.traceContext.Span = spanID
-	span.transactionTimestamp = transactionTimestamp
 	t.spanFramesMinDurationMu.RLock()
 	span.stackFramesMinDuration = t.spanFramesMinDuration
 	t.spanFramesMinDurationMu.RUnlock()
@@ -105,9 +114,6 @@ func (t *Tracer) startSpan(name, spanType string, transactionID SpanID, opts Spa
 	span.parentID = opts.Parent.Span
 	span.transactionID = transactionID
 	span.timestamp = opts.Start
-	if span.timestamp.IsZero() {
-		span.timestamp = time.Now()
-	}
 	return span
 }
 
@@ -118,7 +124,6 @@ type Span struct {
 	parentID               SpanID
 	transactionID          SpanID
 	stackFramesMinDuration time.Duration
-	transactionTimestamp   time.Time
 	timestamp              time.Time
 
 	Name     string
@@ -211,5 +216,15 @@ type SpanOptions struct {
 
 	// Start is the start time of the span. If this has the zero value,
 	// time.Now() will be used instead.
+	//
+	// When a span is created using Transaction.StartSpanOptions, the
+	// span timestamp is internally calculated relative to the transaction
+	// timestamp.
+	//
+	// When Tracer.StartSpan is used, this timestamp should be pre-calculated
+	// as relative from the transaction start time, i.e. by calculating the
+	// time elapsed since the transaction started, and adding that to the
+	// transaction timestamp. Calculating the timstamp in this way will ensure
+	// monotonicity of events within a transaction.
 	Start time.Time
 }